Moved TurnIconHelper and TurnIconResource from libnavui-base to libnavigation-base module.
