+++ libnavui-maneuver/src/main/java/com/mapbox/navigation/ui/maneuver/api/MapboxTurnIconsApi.kt
@@ -4,7 +4,7 @@ import com.mapbox.api.directions.v5.models.ManeuverModifier
 import com.mapbox.bindgen.Expected
 import com.mapbox.bindgen.ExpectedFactory
 import com.mapbox.navigation.base.internal.maneuver.ManeuverTurnIcon
-import com.mapbox.navigation.ui.base.internal.TurnIconHelper
+import com.mapbox.navigation.base.internal.maneuver.TurnIconHelper
 import com.mapbox.navigation.ui.maneuver.model.TurnIconError
 import com.mapbox.navigation.ui.maneuver.model.TurnIconResources
 import com.mapbox.navigation.ui.utils.internal.ifNonNull
