+++ libnavui-maneuver/src/test/java/com/mapbox/navigation/ui/maneuver/TurnIconHelperTest.kt
@@ -19,7 +19,7 @@ import com.mapbox.api.directions.v5.models.StepManeuver.ROUNDABOUT
 import com.mapbox.api.directions.v5.models.StepManeuver.ROUNDABOUT_TURN
 import com.mapbox.api.directions.v5.models.StepManeuver.TURN
 import com.mapbox.navigation.base.internal.maneuver.ManeuverTurnIcon
-import com.mapbox.navigation.ui.base.internal.TurnIconHelper
+import com.mapbox.navigation.base.internal.maneuver.TurnIconHelper
 import com.mapbox.navigation.ui.maneuver.model.TurnIconResources
 import org.junit.Assert.assertEquals
 import org.junit.Test
