+++ libtrip-notification/build.gradle
@@ -33,7 +33,6 @@ dependencies {
 
     implementation dependenciesList.kotlinStdLib
     implementation project(':libnavigation-util')
-    implementation project(':libnavui-base')
 
     implementation dependenciesList.androidXCore
