+++ libtrip-notification/src/test/java/com/mapbox/navigation/trip/notification/internal/NotificationTurnIconHelperTest.kt
@@ -4,7 +4,7 @@ import com.mapbox.api.directions.v5.models.ManeuverModifier
 import com.mapbox.api.directions.v5.models.StepManeuver
 import com.mapbox.navigation.base.internal.maneuver.ManeuverTurnIcon
 import com.mapbox.navigation.trip.notification.R
-import com.mapbox.navigation.ui.base.internal.TurnIconHelper
+import com.mapbox.navigation.base.internal.maneuver.TurnIconHelper
 import org.junit.Assert.assertEquals
 import org.junit.Test
